requirements.txt
django==1.5.1
oauth2==1.5.211
pycountry
python-googlegeocoder
facepy
